move SceneLocationViewDelegate to dedicated file
