+++ ARKit+CoreLocation/Source/SceneLocationView.swift
@@ -11,40 +11,6 @@ import ARKit
 import CoreLocation
 import MapKit
 
-public protocol SceneLocationViewDelegate: class {
-    func sceneLocationViewDidAddSceneLocationEstimate(sceneLocationView: SceneLocationView,
-                                                      position: SCNVector3,
-                                                      location: CLLocation)
-    func sceneLocationViewDidRemoveSceneLocationEstimate(sceneLocationView: SceneLocationView,
-                                                         position: SCNVector3,
-                                                         location: CLLocation)
-
-    ///After a node's location is initially set based on current location,
-    ///it is later confirmed once the user moves far enough away from it.
-    ///This update uses location data collected since the node was placed to give a more accurate location.
-    func sceneLocationViewDidConfirmLocationOfNode(sceneLocationView: SceneLocationView, node: LocationNode)
-
-    func sceneLocationViewDidSetupSceneNode(sceneLocationView: SceneLocationView, sceneNode: SCNNode)
-
-    func sceneLocationViewDidUpdateLocationAndScaleOfLocationNode(sceneLocationView: SceneLocationView,
-                                                                  locationNode: LocationNode)
-}
-
-public extension SceneLocationViewDelegate {
-    func sceneLocationViewDidAddSceneLocationEstimate(sceneLocationView: SceneLocationView,
-                                                      position: SCNVector3,
-                                                      location: CLLocation) { }
-    func sceneLocationViewDidRemoveSceneLocationEstimate(sceneLocationView: SceneLocationView,
-                                                         position: SCNVector3,
-                                                         location: CLLocation) { }
-
-    func sceneLocationViewDidConfirmLocationOfNode(sceneLocationView: SceneLocationView, node: LocationNode) { }
-    func sceneLocationViewDidSetupSceneNode(sceneLocationView: SceneLocationView, sceneNode: SCNNode) { }
-
-    func sceneLocationViewDidUpdateLocationAndScaleOfLocationNode(sceneLocationView: SceneLocationView,
-                                                                  locationNode: LocationNode) { }
-}
-
 ///Different methods which can be used when determining locations (such as the user's location).
 public enum LocationEstimateMethod {
     ///Only uses core location data.
