+++ ARKit+CoreLocation/Source/SceneLocationViewDelegate.swift
@@ -1,0 +1,46 @@
+//
+//  SceneLocationViewDelegate.swift
+//  ARKit+CoreLocation
+//
+//  Created by Ilya Seliverstov on 09/08/2017.
+//  Copyright © 2017 Project Dent. All rights reserved.
+//
+
+import Foundation
+import ARKit
+import CoreLocation
+import MapKit
+
+public protocol SceneLocationViewDelegate: class {
+    func sceneLocationViewDidAddSceneLocationEstimate(sceneLocationView: SceneLocationView,
+                                                      position: SCNVector3,
+                                                      location: CLLocation)
+    func sceneLocationViewDidRemoveSceneLocationEstimate(sceneLocationView: SceneLocationView,
+                                                         position: SCNVector3,
+                                                         location: CLLocation)
+
+    ///After a node's location is initially set based on current location,
+    ///it is later confirmed once the user moves far enough away from it.
+    ///This update uses location data collected since the node was placed to give a more accurate location.
+    func sceneLocationViewDidConfirmLocationOfNode(sceneLocationView: SceneLocationView, node: LocationNode)
+
+    func sceneLocationViewDidSetupSceneNode(sceneLocationView: SceneLocationView, sceneNode: SCNNode)
+
+    func sceneLocationViewDidUpdateLocationAndScaleOfLocationNode(sceneLocationView: SceneLocationView,
+                                                                  locationNode: LocationNode)
+}
+
+public extension SceneLocationViewDelegate {
+    func sceneLocationViewDidAddSceneLocationEstimate(sceneLocationView: SceneLocationView,
+                                                      position: SCNVector3,
+                                                      location: CLLocation) { }
+    func sceneLocationViewDidRemoveSceneLocationEstimate(sceneLocationView: SceneLocationView,
+                                                         position: SCNVector3,
+                                                         location: CLLocation) { }
+
+    func sceneLocationViewDidConfirmLocationOfNode(sceneLocationView: SceneLocationView, node: LocationNode) { }
+    func sceneLocationViewDidSetupSceneNode(sceneLocationView: SceneLocationView, sceneNode: SCNNode) { }
+
+    func sceneLocationViewDidUpdateLocationAndScaleOfLocationNode(sceneLocationView: SceneLocationView,
+                                                                  locationNode: LocationNode) { }
+}
